requirements.txt
streamlit
pandas
numpy
pinecone-client
sentence-transformers
